pyproject.toml
[tool.poetry]
name = "yt-api"
version = "0.1.0"
description = "YouTube API"
authors = ["Kevin H <git@thehollywoods.net>"]
readme = "README.md"
packages = [{include = "yt_api"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-core = "^2.11.1"
google-api-python-client = "^2.93.0"
google-auth = "^2.22.0"
google-auth-oauthlib = "^1.0.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
